+++ tests/testthat/test-gg-plots.R
@@ -33,9 +33,9 @@ test_that("density", {
 
 test_that("cor", {
 
-  expect_warning(
+  expect_error(
     ggally_cor(tips, mapping = ggplot2::aes_string(x = "total_bill", y = "tip"), use = "NOTFOUND"),
-    "correlation 'use' not found"
+    "should be one of"
   )
 
 
@@ -60,9 +60,6 @@ test_that("cor", {
       msg
     )
   }
-  expect_err(corAlignPercent = 0.9, "'corAlignPercent' is deprecated")
-  expect_err(corMethod = "pearson", "'corMethod' is deprecated")
-  expect_err(corUse = "complete.obs", "'corUse' is deprecated")
 
   expect_print(ggally_cor(ti, ggplot2::aes(x = total_bill, y = tip, color = I("green"))))
 
@@ -88,7 +85,7 @@ test_that("cor", {
       ti,
       ggplot2::aes(x = total_bill, y = tip, color = size)
     ),
-    "ggally_cor: mapping color column"
+    "must be categorical"
   )
   expect_silent(
     ggally_cor(
@@ -133,9 +130,9 @@ test_that("dates", {
 
   class(nas) <- c("NOTFOUND", "data.frame")
   p <- ggally_cor(nas, ggplot2::aes(x = date, y = ozone))
-  expect_equal(get("aes_params", envir = p$layers[[1]])$label, "Corr:\n0.278")
+  expect_equal(get("aes_params", envir = p$layers[[1]])$label, "Corr:\n0.278 ***")
   p <- ggally_cor(nas, ggplot2::aes(y = date, x = ozone))
-  expect_equal(get("aes_params", envir = p$layers[[1]])$label, "Corr:\n0.278")
+  expect_equal(get("aes_params", envir = p$layers[[1]])$label, "Corr:\n0.278 ***")
 
   p <- ggally_barDiag(nas, ggplot2::aes(x = date))
   expect_equal(mapping_string(p$mapping$x), "date")