+++ NEWS.md
@@ -1,3 +1,22 @@
+GGally (development version)
+---------------------
+
+
+`ggally_cor`
+* New implementation using `ggally_exec_text()`
+* Will now hide the grid by default and add a border (`displayGrid = FALSE`)
+* Deprecated parameters have been removed
+* Added the ability to display significance stars (`stars = TRUE`)
+* Alignment has been fixed so both short and long names should be displayed within view. `alignPercent` now corresponds to the center of the text.
+* Added the ability to separate the arguments sent to the title and the groups (`title_args` and `group_args`)
+* Digits now represents the total number of digits after the decimal place.
+* To use the old version, change your `ggally_cor` function calls to `ggally_cor_v1_5`.
+
+`ggally_exec_text`
+* New function which is a generialized version of `ggally_cor()`
+* Use this method to create functions similar to `ggally_cor()` that return any text value given and `x` and `y` vector of data
+
+
 GGally 1.5.0
 ---------------------